# faster implementation of sparse/sparse dot product