@vbarrielle 2019-03-14
Fixes #151